+++ apps/admin/src/app/study/[id]/edit/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { PageLayout } from '@hiarc-platform/ui';
+import { StudyEditPage } from '@/features/study/pages';
 import React from 'react';
-import { StudyEditPage } from '@/features/study/pages/study-edit-page/study-edit-page';
 
 export default function EditStudyPage(): React.ReactElement {
   return (
