+++ apps/admin/src/features/study/api/index.ts
@@ -1,5 +1,5 @@
 // Study API exports
 export { studyApi } from './study';
-
+export { studyMemberApi } from './study_member';
 // Re-export all types
 export type * from './study';
