+++ apps/admin/src/features/study/components/index.ts
@@ -1,4 +1,3 @@
-// Study API exports
 export { StudyFormWrapper } from './study-form-wrapper.tsx/StudyFormWrapper';
 export { StudyTable } from './study-table/study-table';
 export { TabSection } from './tab-section/TabSection';
