+++ apps/admin/src/features/study/components/tab-section/dialog/CreateAttendanceCodeDialogWrapper.tsx
@@ -1,5 +1,5 @@
 import { CreateAttendanceCodeDialog } from '@hiarc-platform/ui';
-import { useCreateAttendanceCode } from '../../../hooks';
+import { useCreateAttendanceCode } from '@/features/study/hooks';
 
 interface CreateAttendanceCodeDialogWrapperProps {
   studyId: number;
