+++ apps/admin/src/features/study/hooks/index.ts
@@ -1,14 +1,2 @@
-// Study hooks exports
-export { useStudies } from './use-studies';
-export { useStudy } from './use-study';
-export { useCreateStudy } from './use-create-study';
-export { useUpdateStudy } from './use-update-study';
-export { useLecturesByStudy } from './use-lectures-by-study';
-export { useStudyOptions } from './use-study-options';
-export { useCreateAssignment } from './use-create-assignment';
-export { useAssignment } from './use-assignment';
-export { useAttendanceCode } from './use-attendance-code';
-export { useCreateAttendanceCode } from './use-create-attendance-code';
-export { useStudyAnnouncements } from './use-study-announcements';
-export { useStudyInitialForm } from './use-study-initial-form';
-export { useDeleteLecture } from './use-delete-lecture';
+export * from './study';
+export * from './member';
