+++ apps/admin/src/features/study/hooks/member/mutation/use-create-group.ts
@@ -1,7 +1,7 @@
 import { useMutation, UseMutationResult } from '@tanstack/react-query';
-import { studyApi } from '../api/study';
 import { DialogUtil } from '@hiarc-platform/ui';
 import { CreateGroupRequest } from '@hiarc-platform/shared';
+import { studyMemberApi } from '@/features/study/api';
 
 export function useCreateGroup(): UseMutationResult<
   void,
@@ -11,7 +11,7 @@ export function useCreateGroup(): UseMutationResult<
 > {
   const mutation = useMutation({
     mutationFn: ({ studyId, groupData }: { studyId: number; groupData: CreateGroupRequest }) =>
-      studyApi.CREATE_GROUP(studyId, groupData),
+      studyMemberApi.CREATE_GROUP(studyId, groupData),
     onSuccess: () => {
       DialogUtil.showSuccess('그룹이 성공적으로 생성되었습니다.');
     },
