+++ apps/admin/src/features/study/hooks/member/mutation/use-edit-group.ts
@@ -1,7 +1,7 @@
 import { useMutation, UseMutationResult } from '@tanstack/react-query';
-import { studyApi } from '../api/study';
 import { DialogUtil } from '@hiarc-platform/ui';
 import { CreateGroupRequest } from '@hiarc-platform/shared';
+import { studyMemberApi } from '@/features/study/api';
 
 export function useEditGroup(): UseMutationResult<
   void,
@@ -18,7 +18,7 @@ export function useEditGroup(): UseMutationResult<
       studyId: number;
       groupId: number;
       groupData: CreateGroupRequest;
-    }) => studyApi.PATCH_GROUP(studyId, groupId, groupData),
+    }) => studyMemberApi.PATCH_GROUP(studyId, groupId, groupData),
     onSuccess: () => {
       DialogUtil.showSuccess('그룹이 성공적으로 수정되었습니다.');
     },
