+++ apps/admin/src/features/study/hooks/member/mutation/use-update-member-status.ts
@@ -1,7 +1,7 @@
 import { useMutation, UseMutationResult } from '@tanstack/react-query';
 import { DialogUtil } from '@hiarc-platform/ui';
-import { studyApi } from '../api';
 import { RoundStatus } from '@hiarc-platform/shared/src/types/study/round-status';
+import { studyMemberApi } from '@/features/study/api';
 
 export function useUpdateMemberStatus(): UseMutationResult<
   void,
@@ -18,7 +18,7 @@ export function useUpdateMemberStatus(): UseMutationResult<
       studyId: number;
       memberId: number;
       roundStatuses: RoundStatus[];
-    }) => studyApi.UPDATE_MEMBER_STATUS(studyId, memberId, roundStatuses),
+    }) => studyMemberApi.UPDATE_MEMBER_STATUS(studyId, memberId, roundStatuses),
     onSuccess: () => {
       DialogUtil.showSuccess('상태가 성공적으로 업데이트되었습니다.');
     },
