+++ apps/admin/src/features/study/hooks/member/mutation/use-withdraw-student.ts
@@ -1,6 +1,6 @@
 import { useMutation, UseMutationResult } from '@tanstack/react-query';
 import { DialogUtil } from '@hiarc-platform/ui';
-import { studyApi } from '../api';
+import { studyMemberApi } from '@/features/study/api';
 
 export function useWithdrawStudent(): UseMutationResult<
   void,
@@ -10,7 +10,7 @@ export function useWithdrawStudent(): UseMutationResult<
 > {
   const mutation = useMutation({
     mutationFn: ({ studyId, memberId }: { studyId: number; memberId: number }) =>
-      studyApi.WITHDRAW_STUDENT(studyId, memberId),
+      studyMemberApi.WITHDRAW_STUDENT(studyId, memberId),
     onSuccess: () => {
       DialogUtil.showSuccess('학생이 성공적으로 탈퇴되었습니다.');
     },
