+++ apps/admin/src/features/study/hooks/member/query/use-study-group-list.ts
@@ -1,11 +1,11 @@
 import { useQuery, UseQueryResult, keepPreviousData } from '@tanstack/react-query';
-import { studyApi } from '../api/study';
 import { StudyGroupList } from '@hiarc-platform/shared';
+import { studyMemberApi } from '@/features/study/api';
 
 export function useStudyGroupList(studyId: number): UseQueryResult<StudyGroupList, Error> {
   const query = useQuery({
     queryKey: ['study-group-list', studyId],
-    queryFn: () => studyApi.GET_STUDY_GROUP_LIST(studyId),
+    queryFn: () => studyMemberApi.GET_STUDY_GROUP_LIST(studyId),
     placeholderData: keepPreviousData,
     staleTime: 5 * 60 * 1000, // 5 minutes
     gcTime: 10 * 60 * 1000, // 10 minutes
