+++ apps/admin/src/features/study/hooks/study/mutation/use-delete-lecture.ts
@@ -1,5 +1,5 @@
 import { useMutation, UseMutationResult, useQueryClient } from '@tanstack/react-query';
-import { studyApi } from '../api/study';
+import { studyApi } from '@/features/study/api';
 import { DialogUtil } from '@hiarc-platform/ui';
 
 export function useDeleteLecture(): UseMutationResult<
