+++ apps/admin/src/features/study/hooks/study/mutation/use-update-study.ts
@@ -1,6 +1,6 @@
 import { useMutation, UseMutationResult, useQueryClient } from '@tanstack/react-query';
-import { studyApi } from '../api/study';
-import type { UpdateStudyRequest } from '../api/study';
+import { studyApi } from '@/features/study/api';
+import type { UpdateStudyRequest } from '@/features/study/api';
 import { DialogUtil } from '@hiarc-platform/ui';
 import { useRouter } from 'next/navigation';
