+++ apps/admin/src/features/study/hooks/study/query/use-study-options.ts
@@ -1,5 +1,5 @@
 import { useQuery, UseQueryResult, keepPreviousData } from '@tanstack/react-query';
-import { studyApi } from '../api/study';
+import { studyApi } from '@/features/study/api';
 import { StudySummary, PageableModel, SelectOption } from '@hiarc-platform/shared';
 
 export function useStudyOptions(semesterId?: number | null): UseQueryResult<SelectOption[], Error> {
