+++ apps/admin/src/features/study/pages/index.ts
@@ -1,2 +1,3 @@
 export { StudyCreatePage } from './study-create-page/study-create-page';
+export { StudyEditPage } from './study-edit-page/study-edit-page';
 export { StudyListPage } from './study-list-page/study-list-page';
