--- packages/ui/src/components/study/create-study-form/index.ts
@@ -1,4 +1,0 @@
-export { LabeledMultiSelect } from './labeled-multi-select';
-export { LabeledTimePicker } from './labeled-time-picker';
-export { InformaionSection } from './information-section';
-export { StudyForm } from './study-form';
