+++ packages/ui/src/components/study/index.ts
@@ -1,3 +1,3 @@
 export { StudyAnnouncementTable } from './StudyAnnouncementTable/StudyAnnouncementTable';
 export { StudentList } from './StudentList';
-export { StudyForm } from './create-study-form';
+export { StudyForm } from './study-form';
