+++ packages/ui/src/components/study/study-form/LabeledMultiSelect.tsx
