+++ packages/ui/src/components/study/study-form/LabeledTimePicker.tsx
